% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_data_frame.R
\name{as.data.frame.maze}
\alias{as.data.frame.maze}
\title{Data frame coercion (for plotting)}
\usage{
\method{as.data.frame}{maze}(
  x,
  ...,
  walls = FALSE,
  method = c("geom_point", "geom_path", "geom_tile"),
  start = NULL,
  end = NULL
)
}
\arguments{
\item{x}{A \code{\link[=maze]{maze()}} object.}

\item{...}{Currently ignored.}

\item{walls}{logical value, indicating that the walls of the maze should be
plotted, rather than the paths through the maze. Default is \code{FALSE}.}

\item{method}{Target \code{grid} or \code{ggplot2} plotting method as a string.
Some \code{method} values may only work for either the maze object or its solution.}

\item{start, end}{The coordinates of the starting and ending points of a path
through the maze, or descriptions of relative locations (eg.
\code{"topleft"}, see \code{\link[=find_maze_refpoint]{find_maze_refpoint()}}. If not provided, no path will be drawn.}
}
\value{
A data frame.
}
\description{
\code{\link[=as.data.frame]{as.data.frame()}} methods to coerce \code{\link[=maze]{maze()}} objects
or their solutions into data frames suitable for plotting with
\code{grid} or \code{ggplot2} graphics systems.
}
\examples{
x <- maze(12L, 12L)
dfw <- as.data.frame(x, walls = TRUE)
dfs <- as.data.frame(x, start = "top", end = "bottom")
if (require("ggplot2", quietly = TRUE)) {
   ggplot(dfw, aes(x, y)) + geom_tile() + coord_equal() + theme_void() + 
   geom_path(aes(x=x, y=y), colour = "red", linewidth = 2, data = dfs)
}
}
\seealso{
\code{\link[=plot.maze]{plot.maze()}}, \code{\link[grid:grid.points]{grid::pointsGrob()}}
}
