% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maze.R
\name{as.maze}
\alias{as.maze}
\title{Convert matrix to maze}
\usage{
as.maze(x)
}
\arguments{
\item{x}{A matrix-like object to be made into a maze.}
}
\description{
A function to convert a matrix into a \code{maze} object. Binary matrices can
be used to make mazes in interesting, non-rectangular shapes.
}
\details{
If \code{x} is a binary matrix (or otherwise contains only two
  unique values), the maze will be constrained to occupy only those cells
  containing a 1 (the higher value).
}
\examples{
mat <- matrix(NA, 10, 10)
m <- as.maze(mat)

# circular maze
mat <- matrix(1, 30, 30)
for(i in 1:nrow(mat)){
    for(j in 1:ncol(mat)){
        if((i-15.5)^2+(j-15.5)^2 > 220){
            mat[i,j] <- 0
        }
    }
}
}
