% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced.R
\name{expand_matrix}
\alias{expand_matrix}
\title{Double the dimensions of a matrix}
\usage{
expand_matrix(m)
}
\arguments{
\item{m}{A \code{matrix}.}
}
\value{
A matrix with the same distribution of values as the original, but doubled in
size along both dimensions.
}
\description{
A function to double the size of a matrix, with each cell in the original
becoming four cells in the expanded matrix.
}
\examples{
m <- matrix(1:6, nrow = 2)
expand_matrix(m)

}
