% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maze.R
\name{print.maze}
\alias{print.maze}
\title{Print Maze}
\usage{
\method{print}{maze}(x, ..., walls = FALSE, start = NULL, end = NULL, compress = "v")
}
\arguments{
\item{x}{A \code{maze} object.}

\item{...}{Further arguments passed to
\code{\link[bittermelon]{print.bm_bitmap}}, if the \code{bittermelon}
package is available.}

\item{walls}{logical value, indicating that the walls of the maze should be
plotted, rather than the paths through the maze. Default is \code{FALSE}.}

\item{start, end}{The coordinates of the starting and ending points of a path
through the maze, or descriptions of relative locations (eg.
\code{"topleft"}, see \code{\link{find_maze_refpoint}}). If not provided,
no path will be drawn.}

\item{compress}{How to compress the bitmap representation, see
\code{\link[bittermelon]{print.bm_bitmap}}. For \code{maze} objects, the
default is \code{"v"} for vertical.}
}
\description{
Display a \code{maze} object.
}
\details{
If the \code{\link[bittermelon]{bittermelon}} package is available,
  \code{print.maze} prints a representation of the maze az a bitmap object.
  Otherwise, it prints the matrix representation.
}
\examples{
m <- maze(10,10)
print(m)

}
