% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{solve_maze}
\alias{solve_maze}
\title{Find a path through a maze}
\usage{
solve_maze(maze, start = "bottomleft", end = "topright")
}
\arguments{
\item{maze}{A \code{\link{maze}} object.}

\item{start}{The matrix indices of the starting point. If not provided, this
will be as far to the left as possible (lowest column index) at the bottom
(lowest row index).}

\item{end}{The matrix indices of the end point. If not provided, this will be
as far to the right as possible (highest column index) at the top (highest
row index).}
}
\value{
A \code{matrix} containing the coordinates of the path through the
  maze. Note that the x-coordinate (column index) comes first, so for the
  corresponding indices in the original matrix, these coordinates will need
  to be reversed.
}
\description{
A function that finds the shortest path between points in a maze.
}
\examples{
m <- maze(15,15)
p <- solve_maze(m)

plot(m, walls = TRUE)
lines(p, col = 2, lwd = 3)

}
