% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maze.R
\name{maze}
\alias{maze}
\title{Generate a random maze}
\usage{
maze(nrow, ncol)
}
\arguments{
\item{nrow}{Number of rows.}

\item{ncol}{Number of columns.}
}
\value{
An object of class \code{maze}, which is a subclass of \code{matrix}.
}
\description{
A function that produces a rectangular, procedurally generated random maze of
a given size.
}
\examples{
maze(10,10)
}
